% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainKNPTS.R
\name{predictKNPTS}
\alias{predictKNPTS}
\title{Make predictions using KNPTS (K-Nearest Patterns in Time Series)}
\usage{
predictKNPTS(ts, h, n)
}
\arguments{
\item{n}{}
}
\value{
model
}
\description{
Make predictions using KNPTS (K-Nearest Patterns in Time Series)
}
